String response for api call instead of JSONObject
Change the client api call method to return a String as a response
instead of a JSONObject.  Not all api methods will return a
JSONObject, some will return JSONArray.  Allows flexibility of
resopnse in the future.

Change package name to client versus sdk.  Makes more defined
when adding a components package to the sdk later.