Authenticated services list component.  Extensions.
Created an component that shows the services supported by Singly in
a list with the authenticated services with checkmarks.  Allows users
to authenticate and deauthenticate from a service.

Reworked the SinglyClient class to hold the client id and secret
internally. Added a static getInstance method to return singleton.

Refactored AuthenticationActivity to allow experts to subclass and
hook into the authentication process web page events.